refactor(cli): extract _bootstrap module to break cyclic imports
