+++ src/godspeed/cli.py
@@ -18,6 +18,7 @@ from uuid import uuid4
 import click
 
 from godspeed import __version__
+from godspeed._bootstrap import _build_tool_registry, _load_env_files
 from godspeed.config import DEFAULT_GLOBAL_DIR
 
 
@@ -60,123 +61,6 @@ logger = logging.getLogger(__name__)
 
 OLLAMA_URL = "http://localhost:11434/api/tags"
 OLLAMA_STARTUP_TIMEOUT = 15  # seconds to wait for ollama to come up
-
-
-def _parse_env_file(path: Path) -> dict[str, str]:
-    """Parse a simple ``KEY=value`` env file into a dict.
-
-    Supports:
-    - ``# ...`` comments (line must start with ``#``; inline comments aren't stripped)
-    - Blank lines
-    - Optional surrounding ``"..."`` or ``'...'`` quotes on values
-
-    Malformed lines (no ``=``, empty key) are silently skipped — an env
-    file should never crash startup. Returns an empty dict if the file
-    is missing or unreadable.
-    """
-    result: dict[str, str] = {}
-    if not path.exists():
-        return result
-    try:
-        text = path.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError) as exc:
-        # OSError: permissions / transient FS issue.
-        # UnicodeDecodeError: someone wrote a binary blob or cp1252 file
-        # into .env.local — log and skip rather than crash the CLI.
-        logger.debug("Could not read env file %s: %s", path, exc)
-        return result
-    for raw_line in text.splitlines():
-        line = raw_line.strip()
-        if not line or line.startswith("#"):
-            continue
-        if "=" not in line:
-            continue
-        key, _, value = line.partition("=")
-        key = key.strip()
-        value = value.strip()
-        if not key:
-            continue
-        # Strip a matching pair of surrounding quotes so
-        # ``KEY="value with spaces"`` works as expected.
-        if len(value) >= 2 and value[0] == value[-1] and value[0] in ('"', "'"):
-            value = value[1:-1]
-        result[key] = value
-    return result
-
-
-def _load_env_files(project_dir: Path | None = None) -> list[tuple[Path, list[str]]]:
-    """Load env files from standard locations into ``os.environ``.
-
-    Precedence (highest to lowest):
-
-    1. **Shell environment** — always wins. Keys already in ``os.environ``
-       are never overwritten. This keeps ``$env:NVIDIA_NIM_API_KEY = ...;
-       godspeed`` one-offs working and matches every ``.env`` library's
-       convention.
-    2. ``<project_dir>/.godspeed/.env.local`` — project-scoped override
-    3. ``<project_dir>/.godspeed/.env`` — project-scoped defaults
-    4. ``~/.godspeed/.env.local`` — user-wide override
-    5. ``~/.godspeed/.env`` — user-wide defaults
-
-    Files are merged in low-to-high priority order so higher-priority
-    files overwrite lower-priority ones BEFORE the single shell-env
-    check. In-file semantics mirror dotenv / Vite / Next: ``.env.local``
-    is the gitignored local override, ``.env`` is the checked-in default.
-
-    Returns a list of ``(path, injected_keys)`` pairs for each file
-    that contributed at least one variable — used by the caller to log
-    which files the session picked up. Never raises; bad files are
-    silently skipped so a malformed ``.env.local`` can't brick the CLI.
-    """
-
-    # Lowest priority first so later files overwrite earlier ones in
-    # the merged dict. Shell-env wins is applied at the very end.
-    candidates: list[Path] = [
-        DEFAULT_GLOBAL_DIR / ".env",
-        DEFAULT_GLOBAL_DIR / ".env.local",
-    ]
-    if project_dir is not None:
-        candidates.extend(
-            [
-                project_dir / ".godspeed" / ".env",
-                project_dir / ".godspeed" / ".env.local",
-            ]
-        )
-
-    resolved: dict[str, str] = {}
-    contributions: list[tuple[Path, list[str]]] = []
-    for path in candidates:
-        parsed = _parse_env_file(path)
-        if not parsed:
-            continue
-        # Record which keys this file will contribute to the merged view.
-        # (The actual injection into os.environ happens after the full
-        # merge; shell-env wins over everything below.)
-        contributions.append((path, sorted(parsed.keys())))
-        resolved.update(parsed)
-
-    loaded: list[tuple[Path, list[str]]] = []
-    injected_keys: set[str] = set()
-    for key, value in resolved.items():
-        if key in os.environ:
-            continue
-        os.environ[key] = value
-        injected_keys.add(key)
-
-    # Re-project contributions onto actually-injected keys, so the log
-    # reflects what the session ended up with after shell-env overrides.
-    for path, keys in contributions:
-        effective = [k for k in keys if k in injected_keys]
-        if not effective:
-            continue
-        loaded.append((path, effective))
-        logger.info(
-            "Loaded %d env var(s) from %s: %s",
-            len(effective),
-            path,
-            ", ".join(effective),
-        )
-    return loaded
 
 
 def _setup_logging(verbose: bool) -> None:
@@ -307,137 +191,6 @@ def _ensure_llamacpp(console: Any | None = None) -> bool:
             f"python scripts/setup_qwen36_local.py --build-only[/{WARNING}]"
         )
     return False
-
-
-def _build_tool_registry(tool_set: str = "full") -> tuple:
-    """Create all tool instances and register them.
-
-    Args:
-        tool_set: one of "local", "web", "full" (default). "local" hides
-            web_search/web_fetch/github so weak models don't pick them
-            over file_read/grep_search on local-codebase tasks. "web"
-            adds the web tools back. "full" registers everything.
-
-    Returns:
-        (ToolRegistry, dict[str, RiskLevel]) tuple.
-    """
-    from godspeed.tools.tool_sets import get_allowed_tool_names
-
-    allowed = get_allowed_tool_names(tool_set)
-    from godspeed.tools.base import RiskLevel
-    from godspeed.tools.file_edit import FileEditTool
-    from godspeed.tools.file_move import FileMoveTool
-    from godspeed.tools.file_read import FileReadTool
-    from godspeed.tools.file_write import FileWriteTool
-    from godspeed.tools.registry import ToolRegistry
-
-    registry = ToolRegistry()
-    risk_levels: dict[str, RiskLevel] = {}
-
-    from godspeed.tools.background import BackgroundCheckTool
-    from godspeed.tools.complexity import ComplexityTool
-    from godspeed.tools.coverage import CoverageTool
-    from godspeed.tools.db_query import DbQueryTool
-    from godspeed.tools.dep_audit import DepAuditTool
-    from godspeed.tools.generate_tests import GenerateTestsTool
-    from godspeed.tools.git import GitTool
-    from godspeed.tools.glob_search import GlobSearchTool
-    from godspeed.tools.grep_search import GrepSearchTool
-    from godspeed.tools.notebook import NotebookEditTool
-    from godspeed.tools.repo_map import RepoMapTool
-    from godspeed.tools.security_scan import SecurityScanTool
-    from godspeed.tools.shell import ShellTool
-    from godspeed.tools.test_runner import TestRunnerTool
-    from godspeed.tools.traceback_analyzer import TracebackAnalyzerTool
-    from godspeed.tools.verify import VerifyTool
-    from godspeed.tools.web_fetch import WebFetchTool
-    from godspeed.tools.web_search import WebSearchTool
-
-    try:
-        from godspeed.tools.stock_price import StockPriceTool
-
-        _stock_price_available = True
-    except ImportError:
-        _stock_price_available = False
-
-    tools: list = [
-        FileReadTool(),
-        TracebackAnalyzerTool(),
-        FileWriteTool(),
-        FileEditTool(),
-        FileMoveTool(),
-        ShellTool(),
-        GlobSearchTool(),
-        GrepSearchTool(),
-        GitTool(),
-        VerifyTool(),
-        RepoMapTool(),
-        TestRunnerTool(),
-        CoverageTool(),
-        SecurityScanTool(),
-        ComplexityTool(),
-        DepAuditTool(),
-        GenerateTestsTool(),
-        WebSearchTool(),
-        WebFetchTool(),
-        NotebookEditTool(),
-        BackgroundCheckTool(),
-        DbQueryTool(),
-    ]
-
-    if _stock_price_available:
-        tools.append(StockPriceTool())
-
-    # Optional tools — register if their dependencies are available
-    try:
-        from godspeed.tools.image_read import ImageReadTool
-
-        tools.append(ImageReadTool())
-    except ImportError:
-        logger.debug("image_read not available")
-
-    try:
-        from godspeed.tools.pdf_read import PdfReadTool
-
-        tools.append(PdfReadTool())
-    except ImportError:
-        logger.debug("pdf_read not available")
-
-    try:
-        from godspeed.tools.github import GithubTool
-
-        tools.append(GithubTool())
-    except ImportError:
-        logger.debug("github not available")
-
-    try:
-        from godspeed.tools.diff_apply import DiffApplyTool
-
-        tools.append(DiffApplyTool())
-    except ImportError:
-        logger.debug("diff_apply not available")
-
-    try:
-        from godspeed.tools.ollama_manager import OllamaTool
-
-        tools.append(OllamaTool())
-    except ImportError:
-        logger.debug("ollama_manager not available")
-
-    try:
-        from godspeed.tools.llamacpp_manager import LlamaCppTool
-
-        tools.append(LlamaCppTool())
-    except ImportError:
-        logger.debug("llamacpp_manager not available")
-
-    for tool in tools:
-        if allowed is not None and tool.name not in allowed:
-            continue
-        registry.register(tool)
-        risk_levels[tool.name] = tool.risk_level
-
-    return registry, risk_levels
 
 
 async def _run_app(
