+++ src/godspeed/mcp_server/server.py
@@ -18,6 +18,7 @@ from mcp.server import NotificationOptions, Server
 from mcp.server.models import InitializationOptions
 from mcp.server.stdio import stdio_server
 
+from godspeed._bootstrap import _build_tool_registry, _load_env_files
 from godspeed.audit.trail import AuditTrail
 from godspeed.config import GodspeedSettings
 from godspeed.mcp_server.schemas import build_mcp_tools
@@ -31,8 +32,6 @@ logger = logging.getLogger(__name__)
 
 def _load_settings_with_optional_config(config_path: Path | None) -> GodspeedSettings:
     """Load settings using standard CLI behavior, with optional file override."""
-    from godspeed.cli import _load_env_files
-
     _load_env_files(project_dir=Path("."))
     if config_path is None:
         return GodspeedSettings()
@@ -86,8 +85,6 @@ class GodspeedMCPServer:
         self.settings = _load_settings_with_optional_config(config_path)
         self.project_dir = Path(self.settings.project_dir).resolve()
         self.session_id = str(uuid4())
-
-        from godspeed.cli import _build_tool_registry
 
         registry, risk_levels = _build_tool_registry()
         task_store = TaskStore()
